app.js 정리 후 navigator 폴더 생성
